introduce useAsync hook
I've noticed that we've started to make unpolled calls to the api. We
use the `Loadable` monad to handle the state of the call during render,
and we are making the call in a side effect, but I've noticed that there
are a few potholes that users can fall into:

* api calls are not guaranteed to return in order, so simply awaiting the value
  and setting it to the state might cause an older call to overwrite a newer
  one.

* `useEffect` expects a function returning undefined or a function that itself
  returns undefined. We prefer to write our async code using `async`/`await` as
  opposed to promises. this means the effect will either need a callback (not
  ideal -- adds a layer of indirection because the effect is now dependent on
  the callback which has its own dependencies) or an IIFE (kinda gross looking,
  introduces a new scope).

To improve the ergonomics of this pattern, this pr introduces a `useAsync` hook.
It calls an async callback as an when its deps change. The return value of the
hook is stored as a Loadable of the return type. To block setting the state when
handling an error, the function can return `NotLoaded`